+++ saas_portal_sale/models/__init__.py
@@ -1,2 +1,4 @@
 from . import product_template
-from . import saas_portal
+from . import sale_order
+from . import saas_portal_plan
+from . import saas_portal_client