+++ saas_portal_sale/doc/changelog.rst
@@ -1,3 +1,9 @@
+`1.1.0`
+------------
+
+- Complete UX revamping
+- Adding of topup products
+
 `1.0.0`
 ------------